backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic>=2.8,<3
pydantic-settings>=2.3,<3
motor>=3.6,<4
pymongo>=4.7,<5
python-multipart>=0.0.9,<0.1
orjson>=3.10,<4
ultralytics>=8.3.0
pillow>=10.4,<11
google-cloud-pubsub>=2.22,<3
google-cloud-storage>=2.18,<3
google-auth>=2.33,<3
tenacity>=9.0,<10
httpx>=0.27,<0.28
structlog>=24.1.0
redis>=6.4.0
